requirements.txt
branca==0.8.0
certifi==2024.8.30
charset-normalizer==3.4.0
folium==0.18.0
geographiclib==2.0
geopandas==1.0.1
geopy==2.4.1
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
kafka-python==2.0.2
MarkupSafe==3.0.2
networkx==3.4.2
numpy==2.1.3
osmnx==2.0.0
packaging==24.2
pandas==2.2.3
pyogrio==0.10.0
pyproj==3.7.0
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
shapely==2.0.6
six==1.16.0
threadpoolctl==3.5.0
tzdata==2024.2
urllib3==2.2.3
xyzservices==2024.9.0
